Move pooling layers to their own file
