+++ keras/layers/convolutional.py
@@ -4,17 +4,11 @@ from __future__ import absolute_import
 from .. import backend as K
 from .. import activations, initializations, regularizers, constraints
 from ..engine import Layer, InputSpec
+from ..utils.np_utils import conv_output_length
 
-
-def conv_output_length(input_length, filter_size, border_mode, stride):
-    if input_length is None:
-        return None
-    assert border_mode in {'same', 'valid'}
-    if border_mode == 'same':
-        output_length = input_length
-    elif border_mode == 'valid':
-        output_length = input_length - filter_size + 1
-    return (output_length + stride - 1) // stride
+# imports for backwards namespace compatibility
+from .pooling import AveragePooling1D, AveragePooling2D, AveragePooling3D
+from .pooling import MaxPooling1D, MaxPooling2D, MaxPooling3D
 
 
 class Convolution1D(Layer):
@@ -390,8 +384,6 @@ class Convolution3D(Layer):
     (tuple of integers, does not include the sample axis),
     e.g. `input_shape=(3, 10, 128, 128)` for 10 frames of 128x128 RGB pictures.
 
-    Note: this layer will only work with Theano for the time being.
-
     # Arguments
         nb_filter: Number of convolution filters to use.
         kernel_dim1: Length of the first dimension in the convolution kernel.
@@ -449,9 +441,6 @@ class Convolution3D(Layer):
                  W_regularizer=None, b_regularizer=None, activity_regularizer=None,
                  W_constraint=None, b_constraint=None,
                  bias=True, **kwargs):
-        if K._BACKEND != 'theano':
-            raise Exception(self.__class__.__name__ +
-                            ' is currently only working with Theano backend.')
         if border_mode not in {'valid', 'same'}:
             raise Exception('Invalid border mode for Convolution3D:', border_mode)
         self.nb_filter = nb_filter
@@ -584,401 +573,6 @@ class Convolution3D(Layer):
                   'bias': self.bias}
         base_config = super(Convolution3D, self).get_config()
         return dict(list(base_config.items()) + list(config.items()))
-
-
-class _Pooling1D(Layer):
-    '''Abstract class for different pooling 1D layers.
-    '''
-    input_dim = 3
-
-    def __init__(self, pool_length=2, stride=None,
-                 border_mode='valid', **kwargs):
-        super(_Pooling1D, self).__init__(**kwargs)
-        if stride is None:
-            stride = pool_length
-        self.pool_length = pool_length
-        self.stride = stride
-        self.st = (self.stride, 1)
-        self.pool_size = (pool_length, 1)
-        assert border_mode in {'valid', 'same'}, 'border_mode must be in {valid, same}'
-        self.border_mode = border_mode
-        self.input_spec = [InputSpec(ndim=3)]
-
-    def get_output_shape_for(self, input_shape):
-        length = conv_output_length(input_shape[1], self.pool_length,
-                                    self.border_mode, self.stride)
-        return (input_shape[0], length, input_shape[2])
-
-    def _pooling_function(self, back_end, inputs, pool_size, strides,
-                          border_mode, dim_ordering):
-        raise NotImplementedError
-
-    def call(self, x, mask=None):
-        x = K.expand_dims(x, -1)   # add dummy last dimension
-        x = K.permute_dimensions(x, (0, 2, 1, 3))
-        output = self._pooling_function(inputs=x, pool_size=self.pool_size,
-                                        strides=self.st,
-                                        border_mode=self.border_mode,
-                                        dim_ordering='th')
-        output = K.permute_dimensions(output, (0, 2, 1, 3))
-        return K.squeeze(output, 3)  # remove dummy last dimension
-
-    def get_config(self):
-        config = {'stride': self.stride,
-                  'pool_length': self.pool_length,
-                  'border_mode': self.border_mode}
-        base_config = super(_Pooling1D, self).get_config()
-        return dict(list(base_config.items()) + list(config.items()))
-
-
-class MaxPooling1D(_Pooling1D):
-    '''Max pooling operation for temporal data.
-
-    # Input shape
-        3D tensor with shape: `(samples, steps, features)`.
-
-    # Output shape
-        3D tensor with shape: `(samples, downsampled_steps, features)`.
-
-    # Arguments
-        pool_length: factor by which to downscale. 2 will halve the input.
-        stride: integer or None. Stride value.
-        border_mode: 'valid' or 'same'.
-            Note: 'same' will only work with TensorFlow for the time being.
-    '''
-
-    def __init__(self, pool_length=2, stride=None,
-                 border_mode='valid', **kwargs):
-        super(MaxPooling1D, self).__init__(pool_length, stride,
-                                           border_mode, **kwargs)
-
-    def _pooling_function(self, inputs, pool_size, strides,
-                          border_mode, dim_ordering):
-        output = K.pool2d(inputs, pool_size, strides,
-                          border_mode, dim_ordering, pool_mode='max')
-        return output
-
-
-class AveragePooling1D(_Pooling1D):
-    '''Average pooling for temporal data.
-
-    # Arguments
-        pool_length: factor by which to downscale. 2 will halve the input.
-        stride: integer or None. Stride value.
-        border_mode: 'valid' or 'same'.
-            Note: 'same' will only work with TensorFlow for the time being.
-
-    # Input shape
-        3D tensor with shape: `(samples, steps, features)`.
-
-    # Output shape
-        3D tensor with shape: `(samples, downsampled_steps, features)`.
-    '''
-
-    def __init__(self, pool_length=2, stride=None,
-                 border_mode='valid', **kwargs):
-        super(AveragePooling1D, self).__init__(pool_length, stride,
-                                               border_mode, **kwargs)
-
-    def _pooling_function(self, inputs, pool_size, strides,
-                          border_mode, dim_ordering):
-        output = K.pool2d(inputs, pool_size, strides,
-                          border_mode, dim_ordering, pool_mode='avg')
-        return output
-
-
-class _Pooling2D(Layer):
-    '''Abstract class for different pooling 2D layers.
-    '''
-
-    def __init__(self, pool_size=(2, 2), strides=None, border_mode='valid',
-                 dim_ordering=K.image_dim_ordering(), **kwargs):
-        super(_Pooling2D, self).__init__(**kwargs)
-        self.pool_size = tuple(pool_size)
-        if strides is None:
-            strides = self.pool_size
-        self.strides = tuple(strides)
-        assert border_mode in {'valid', 'same'}, 'border_mode must be in {valid, same}'
-        self.border_mode = border_mode
-        assert dim_ordering in {'tf', 'th'}, 'dim_ordering must be in {tf, th}'
-        self.dim_ordering = dim_ordering
-        self.input_spec = [InputSpec(ndim=4)]
-
-    def get_output_shape_for(self, input_shape):
-        if self.dim_ordering == 'th':
-            rows = input_shape[2]
-            cols = input_shape[3]
-        elif self.dim_ordering == 'tf':
-            rows = input_shape[1]
-            cols = input_shape[2]
-        else:
-            raise Exception('Invalid dim_ordering: ' + self.dim_ordering)
-
-        rows = conv_output_length(rows, self.pool_size[0],
-                                  self.border_mode, self.strides[0])
-        cols = conv_output_length(cols, self.pool_size[1],
-                                  self.border_mode, self.strides[1])
-
-        if self.dim_ordering == 'th':
-            return (input_shape[0], input_shape[1], rows, cols)
-        elif self.dim_ordering == 'tf':
-            return (input_shape[0], rows, cols, input_shape[3])
-        else:
-            raise Exception('Invalid dim_ordering: ' + self.dim_ordering)
-
-    def _pooling_function(self, inputs, pool_size, strides,
-                          border_mode, dim_ordering):
-        raise NotImplementedError
-
-    def call(self, x, mask=None):
-        output = self._pooling_function(inputs=x, pool_size=self.pool_size,
-                                        strides=self.strides,
-                                        border_mode=self.border_mode,
-                                        dim_ordering=self.dim_ordering)
-        return output
-
-    def get_config(self):
-        config = {'pool_size': self.pool_size,
-                  'border_mode': self.border_mode,
-                  'strides': self.strides,
-                  'dim_ordering': self.dim_ordering}
-        base_config = super(_Pooling2D, self).get_config()
-        return dict(list(base_config.items()) + list(config.items()))
-
-
-class MaxPooling2D(_Pooling2D):
-    '''Max pooling operation for spatial data.
-
-    # Arguments
-        pool_size: tuple of 2 integers,
-            factors by which to downscale (vertical, horizontal).
-            (2, 2) will halve the image in each dimension.
-        strides: tuple of 2 integers, or None. Strides values.
-        border_mode: 'valid' or 'same'.
-            Note: 'same' will only work with TensorFlow for the time being.
-        dim_ordering: 'th' or 'tf'. In 'th' mode, the channels dimension
-            (the depth) is at index 1, in 'tf' mode is it at index 3.
-            It defaults to the `image_dim_ordering` value found in your
-            Keras config file at `~/.keras/keras.json`.
-            If you never set it, then it will be "th".
-
-    # Input shape
-        4D tensor with shape:
-        `(samples, channels, rows, cols)` if dim_ordering='th'
-        or 4D tensor with shape:
-        `(samples, rows, cols, channels)` if dim_ordering='tf'.
-
-    # Output shape
-        4D tensor with shape:
-        `(nb_samples, channels, pooled_rows, pooled_cols)` if dim_ordering='th'
-        or 4D tensor with shape:
-        `(samples, pooled_rows, pooled_cols, channels)` if dim_ordering='tf'.
-    '''
-
-    def __init__(self, pool_size=(2, 2), strides=None, border_mode='valid',
-                 dim_ordering=K.image_dim_ordering(), **kwargs):
-        super(MaxPooling2D, self).__init__(pool_size, strides, border_mode,
-                                           dim_ordering, **kwargs)
-
-    def _pooling_function(self, inputs, pool_size, strides,
-                          border_mode, dim_ordering):
-        output = K.pool2d(inputs, pool_size, strides,
-                          border_mode, dim_ordering, pool_mode='max')
-        return output
-
-
-class AveragePooling2D(_Pooling2D):
-    '''Average pooling operation for spatial data.
-
-    # Arguments
-        pool_size: tuple of 2 integers,
-            factors by which to downscale (vertical, horizontal).
-            (2, 2) will halve the image in each dimension.
-        strides: tuple of 2 integers, or None. Strides values.
-        border_mode: 'valid' or 'same'.
-            Note: 'same' will only work with TensorFlow for the time being.
-        dim_ordering: 'th' or 'tf'. In 'th' mode, the channels dimension
-            (the depth) is at index 1, in 'tf' mode is it at index 3.
-            It defaults to the `image_dim_ordering` value found in your
-            Keras config file at `~/.keras/keras.json`.
-            If you never set it, then it will be "th".
-
-    # Input shape
-        4D tensor with shape:
-        `(samples, channels, rows, cols)` if dim_ordering='th'
-        or 4D tensor with shape:
-        `(samples, rows, cols, channels)` if dim_ordering='tf'.
-
-    # Output shape
-        4D tensor with shape:
-        `(nb_samples, channels, pooled_rows, pooled_cols)` if dim_ordering='th'
-        or 4D tensor with shape:
-        `(samples, pooled_rows, pooled_cols, channels)` if dim_ordering='tf'.
-    '''
-
-    def __init__(self, pool_size=(2, 2), strides=None, border_mode='valid',
-                 dim_ordering=K.image_dim_ordering(), **kwargs):
-        super(AveragePooling2D, self).__init__(pool_size, strides, border_mode,
-                                               dim_ordering, **kwargs)
-
-    def _pooling_function(self, inputs, pool_size, strides,
-                          border_mode, dim_ordering):
-        output = K.pool2d(inputs, pool_size, strides,
-                          border_mode, dim_ordering, pool_mode='avg')
-        return output
-
-
-class _Pooling3D(Layer):
-    '''Abstract class for different pooling 3D layers.
-    '''
-
-    def __init__(self, pool_size=(2, 2, 2), strides=None, border_mode='valid',
-                 dim_ordering=K.image_dim_ordering(), **kwargs):
-        super(_Pooling3D, self).__init__(**kwargs)
-        self.pool_size = tuple(pool_size)
-        if strides is None:
-            strides = self.pool_size
-        self.strides = tuple(strides)
-        assert border_mode in {'valid', 'same'}, 'border_mode must be in {valid, same}'
-        self.border_mode = border_mode
-        assert dim_ordering in {'tf', 'th'}, 'dim_ordering must be in {tf, th}'
-        self.dim_ordering = dim_ordering
-        self.input_spec = [InputSpec(ndim=5)]
-
-    def get_output_shape_for(self, input_shape):
-        if self.dim_ordering == 'th':
-            len_dim1 = input_shape[2]
-            len_dim2 = input_shape[3]
-            len_dim3 = input_shape[4]
-        elif self.dim_ordering == 'tf':
-            len_dim1 = input_shape[1]
-            len_dim2 = input_shape[2]
-            len_dim3 = input_shape[3]
-        else:
-            raise Exception('Invalid dim_ordering: ' + self.dim_ordering)
-
-        len_dim1 = conv_output_length(len_dim1, self.pool_size[0],
-                                      self.border_mode, self.strides[0])
-        len_dim2 = conv_output_length(len_dim2, self.pool_size[1],
-                                      self.border_mode, self.strides[1])
-        len_dim3 = conv_output_length(len_dim3, self.pool_size[2],
-                                      self.border_mode, self.strides[2])
-
-        if self.dim_ordering == 'th':
-            return (input_shape[0], input_shape[1], len_dim1, len_dim2, len_dim3)
-        elif self.dim_ordering == 'tf':
-            return (input_shape[0], len_dim1, len_dim2, len_dim3, input_shape[4])
-        else:
-            raise Exception('Invalid dim_ordering: ' + self.dim_ordering)
-
-    def _pooling_function(self, inputs, pool_size, strides,
-                          border_mode, dim_ordering):
-        raise NotImplementedError
-
-    def call(self, x, mask=None):
-        output = self._pooling_function(inputs=x, pool_size=self.pool_size,
-                                        strides=self.strides,
-                                        border_mode=self.border_mode,
-                                        dim_ordering=self.dim_ordering)
-        return output
-
-    def get_config(self):
-        config = {'pool_size': self.pool_size,
-                  'border_mode': self.border_mode,
-                  'strides': self.strides,
-                  'dim_ordering': self.dim_ordering}
-        base_config = super(_Pooling3D, self).get_config()
-        return dict(list(base_config.items()) + list(config.items()))
-
-
-class MaxPooling3D(_Pooling3D):
-    '''Max pooling operation for 3D data (spatial or spatio-temporal).
-
-    Note: this layer will only work with Theano for the time being.
-
-    # Arguments
-        pool_size: tuple of 3 integers,
-            factors by which to downscale (dim1, dim2, dim3).
-            (2, 2, 2) will halve the size of the 3D input in each dimension.
-        strides: tuple of 3 integers, or None. Strides values.
-        border_mode: 'valid' or 'same'.
-        dim_ordering: 'th' or 'tf'. In 'th' mode, the channels dimension
-            (the depth) is at index 1, in 'tf' mode is it at index 4.
-            It defaults to the `image_dim_ordering` value found in your
-            Keras config file at `~/.keras/keras.json`.
-            If you never set it, then it will be "th".
-
-    # Input shape
-        5D tensor with shape:
-        `(samples, channels, len_pool_dim1, len_pool_dim2, len_pool_dim3)` if dim_ordering='th'
-        or 5D tensor with shape:
-        `(samples, len_pool_dim1, len_pool_dim2, len_pool_dim3, channels)` if dim_ordering='tf'.
-
-    # Output shape
-        5D tensor with shape:
-        `(nb_samples, channels, pooled_dim1, pooled_dim2, pooled_dim3)` if dim_ordering='th'
-        or 5D tensor with shape:
-        `(samples, pooled_dim1, pooled_dim2, pooled_dim3, channels)` if dim_ordering='tf'.
-    '''
-
-    def __init__(self, pool_size=(2, 2, 2), strides=None, border_mode='valid',
-                 dim_ordering=K.image_dim_ordering(), **kwargs):
-        if K._BACKEND != 'theano':
-            raise Exception(self.__class__.__name__ +
-                            ' is currently only working with Theano backend.')
-        super(MaxPooling3D, self).__init__(pool_size, strides, border_mode,
-                                           dim_ordering, **kwargs)
-
-    def _pooling_function(self, inputs, pool_size, strides,
-                          border_mode, dim_ordering):
-        output = K.pool3d(inputs, pool_size, strides,
-                          border_mode, dim_ordering, pool_mode='max')
-        return output
-
-
-class AveragePooling3D(_Pooling3D):
-    '''Average pooling operation for 3D data (spatial or spatio-temporal).
-
-    Note: this layer will only work with Theano for the time being.
-
-    # Arguments
-        pool_size: tuple of 3 integers,
-            factors by which to downscale (dim1, dim2, dim3).
-            (2, 2, 2) will halve the size of the 3D input in each dimension.
-        strides: tuple of 3 integers, or None. Strides values.
-        border_mode: 'valid' or 'same'.
-        dim_ordering: 'th' or 'tf'. In 'th' mode, the channels dimension
-            (the depth) is at index 1, in 'tf' mode is it at index 4.
-            It defaults to the `image_dim_ordering` value found in your
-            Keras config file at `~/.keras/keras.json`.
-            If you never set it, then it will be "th".
-
-    # Input shape
-        5D tensor with shape:
-        `(samples, channels, len_pool_dim1, len_pool_dim2, len_pool_dim3)` if dim_ordering='th'
-        or 5D tensor with shape:
-        `(samples, len_pool_dim1, len_pool_dim2, len_pool_dim3, channels)` if dim_ordering='tf'.
-
-    # Output shape
-        5D tensor with shape:
-        `(nb_samples, channels, pooled_dim1, pooled_dim2, pooled_dim3)` if dim_ordering='th'
-        or 5D tensor with shape:
-        `(samples, pooled_dim1, pooled_dim2, pooled_dim3, channels)` if dim_ordering='tf'.
-    '''
-
-    def __init__(self, pool_size=(2, 2, 2), strides=None, border_mode='valid',
-                 dim_ordering=K.image_dim_ordering(), **kwargs):
-        if K._BACKEND != 'theano':
-            raise Exception(self.__class__.__name__ +
-                            ' is currently only working with Theano backend.')
-        super(AveragePooling3D, self).__init__(pool_size, strides, border_mode,
-                                               dim_ordering, **kwargs)
-
-    def _pooling_function(self, inputs, pool_size, strides,
-                          border_mode, dim_ordering):
-        output = K.pool3d(inputs, pool_size, strides,
-                          border_mode, dim_ordering, pool_mode='avg')
-        return output
 
 
 class UpSampling1D(Layer):
